% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conswat.R
\name{conswat}
\alias{conswat}
\title{Prioritising water bodies for site-based conservation}
\usage{
conswat(data, cols, weights, area, spp, params, n = 7)
}
\arguments{
\item{vector}{}
}
\value{
vector
}
\description{
Basooma et al. 2020. Prioritising the selection of a waterbody
for conservation based on the fish species diversity, composition, and
conservation in Uganda.
}
\details{
The priority index is a novel conservation metric which weights
the waterbody depending on the species richness and rareness among the
water bodies considered. In addition, the species International Union
Conservation for Nature status is weighted where a critically endangered
is weighted 5, endangered = 4, vulnerable = 3, near threatened = 2,
least concern = 1. Based on the assumption by IUCN that both data deficiecnt
and not evaluted should considered threatened until thier status is
established, all their weights was 5. The surface area of the waterbody
 considerd in the computation.
}
\examples{

}
