% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conwatplot.R
\name{conwatplot}
\alias{conwatplot}
\title{BarPlots with ggplot2}
\usage{
conwatplot(data, x, y)
}
\arguments{
\item{data}{Dataset for analysis}

\item{x}{values on the x axis}

\item{y}{values on y the Y-axis}
}
\value{
plot
}
\description{
Making plots for conservation priority index outputs for waterbodies
}
\examples{
conwatplot(gbif, waterbody, Area)

}
